refactor: removing unnecessary archives and adjusting methods
