+++ lib/shared/utils/currency_formater.dart
@@ -1,7 +1,4 @@
-import 'dart:io';
 
 import 'package:intl/intl.dart';
 
-NumberFormat currencyFormatter = NumberFormat.simpleCurrency(
-  locale: Platform.localeName == 'pt_BR' ? 'pt_BR' : 'en_US',
-);
+NumberFormat currencyFormatter = NumberFormat.simpleCurrency(locale: 'pt_BR');
