+++ test/widget_test/details/details_screen_with_provider_test.dart
@@ -1,15 +1,14 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:network_image_mock/network_image_mock.dart';
 import 'package:projeto_criptos/details/view/details_screen.dart';
-import 'package:projeto_criptos/details/widgets/body_details_screen.dart';
+import 'package:projeto_criptos/shared/templates/warren_button.dart';
 import 'package:projeto_criptos/shared/utils/decimal_parse.dart';
 
 import '../../setup/common_asset.dart';
 import '../../setup/setup_widget_tester_override_providers.dart';
 
 void main() {
-  testWidgets(
-      'WHEN load PortfolioScreen THEN ensure returned BodyPortfolioScreen',
+  testWidgets('WHEN load PortfolioScreen THEN press warren button',
       (WidgetTester tester) async {
     mockNetworkImagesFor(() async {
       await loadPageWithProvider(
@@ -23,7 +22,9 @@ void main() {
 
       await tester.pumpAndSettle();
 
-      expect(find.byType(BodyDetailsScreen), findsOneWidget);
+      await tester.tap(find.byType(WarrenButton));
+
+      await tester.pumpAndSettle();
     });
   });
 }
