+++ test/widget_test/l10n/l10n_spanish_test.dart
@@ -5,7 +5,7 @@ import '../../setup/setup_l10n_tester.dart';
 import '../../setup/setup_widget_tester.dart';
 
 void main() {
-  testWidgets('WHEN load L10nWidgetTester THEN test portuguese texts',
+  testWidgets('WHEN load L10nWidgetTester THEN test spanish texts',
       (WidgetTester tester) async {
     await loadPage(tester, const L10nWidgetTester(locale: Locale('es', '')));
