+++ test/widget_test/portfolio/visibility_off_container_test.dart
@@ -4,7 +4,7 @@ import 'package:projeto_criptos/portfolio/widgets/visibility_off_container.dart'
 import '../../setup/setup_widget_tester.dart';
 
 void main() {
-  testWidgets('WHEN load VisibilityOffContainer THEN return its dimensions', (WidgetTester tester) async {
+  testWidgets('WHEN load VisibilityOffContainer THEN find its dimensions', (WidgetTester tester) async {
     await loadPage(tester, const VisibilityOffContainer(witdh: 50, height: 100));
 
     await tester.pumpAndSettle();
